finale/requirements.txt
flask
python-dotenv
requests
langchain_openai
langchain_community
openai